dando permissão pra pipeline android.yml
