+++ mobile_hivemind/app/src/main/AndroidManifest.xml
@@ -21,10 +21,10 @@
         android:theme="@style/Theme.Mobile_hivemind"
         tools:targetApi="31">
         <activity
-            android:name=".ui.auth.CadastroActivity"
+            android:name=".auth.CadastroActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.auth.login.LoginActivity"
+            android:name=".auth.login.LoginActivity"
             android:exported="false" />
         <activity
             android:name=".SplashScreenActivity"
