+++ mobile_hivemind/app/src/main/java/com/aula/mobile_hivemind/auth/CadastroActivity.java
@@ -1,4 +1,4 @@
-package com.aula.mobile_hivemind.ui.auth;
+package com.aula.mobile_hivemind.auth;
 
 import android.os.Bundle;
