+++ mobile_hivemind/app/src/main/java/com/aula/mobile_hivemind/auth/cadastro/CadastroBasicoFragment.java
@@ -1,19 +1,21 @@
-package com.aula.mobile_hivemind.ui.auth.cadastro;
+package com.aula.mobile_hivemind.auth.cadastro;
 
 import android.os.Bundle;
+
 import androidx.fragment.app.Fragment;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
 import com.aula.mobile_hivemind.R;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link CadastroEstruturaFragment#newInstance} factory method to
+ * Use the {@link CadastroBasicoFragment#newInstance} factory method to
  * create an instance of this fragment.
- *
  */
-public class CadastroEstruturaFragment extends Fragment {
+public class CadastroBasicoFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -24,26 +26,26 @@ public class CadastroEstruturaFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
+    public CadastroBasicoFragment() {
+        // Required empty public constructor
+    }
+
     /**
      * Use this factory method to create a new instance of
      * this fragment using the provided parameters.
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment CadastroEstruturaFragment.
+     * @return A new instance of fragment CadastroBasicoFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static CadastroEstruturaFragment newInstance(String param1, String param2) {
-        CadastroEstruturaFragment fragment = new CadastroEstruturaFragment();
+    public static CadastroBasicoFragment newInstance(String param1, String param2) {
+        CadastroBasicoFragment fragment = new CadastroBasicoFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
         fragment.setArguments(args);
         return fragment;
-    }
-
-    public CadastroEstruturaFragment() {
-        // Required empty public constructor
     }
 
     @Override
@@ -59,6 +61,6 @@ public class CadastroEstruturaFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_cadastro_estrutura, container, false);
+        return inflater.inflate(R.layout.fragment_cadastro_basico, container, false);
     }
 }
