+++ mobile_hivemind/app/src/main/java/com/aula/mobile_hivemind/auth/cadastro/CriacaoContaFragment.java
@@ -1,19 +1,21 @@
-package com.aula.mobile_hivemind.ui.auth.cadastro;
+package com.aula.mobile_hivemind.auth.cadastro;
 
 import android.os.Bundle;
+
 import androidx.fragment.app.Fragment;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
 import com.aula.mobile_hivemind.R;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link CadastroPessoalFragment#newInstance} factory method to
+ * Use the {@link CriacaoContaFragment#newInstance} factory method to
  * create an instance of this fragment.
- *
  */
-public class CadastroPessoalFragment extends Fragment {
+public class CriacaoContaFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -24,26 +26,26 @@ public class CadastroPessoalFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
+    public CriacaoContaFragment() {
+        // Required empty public constructor
+    }
+
     /**
      * Use this factory method to create a new instance of
      * this fragment using the provided parameters.
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment CadastroPessoalFragment.
+     * @return A new instance of fragment CriacaoContaFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static CadastroPessoalFragment newInstance(String param1, String param2) {
-        CadastroPessoalFragment fragment = new CadastroPessoalFragment();
+    public static CriacaoContaFragment newInstance(String param1, String param2) {
+        CriacaoContaFragment fragment = new CriacaoContaFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
         fragment.setArguments(args);
         return fragment;
-    }
-
-    public CadastroPessoalFragment() {
-        // Required empty public constructor
     }
 
     @Override
@@ -59,6 +61,6 @@ public class CadastroPessoalFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_cadastro_pessoal, container, false);
+        return inflater.inflate(R.layout.fragment_criacao_conta, container, false);
     }
 }
