+++ mobile_hivemind/app/src/main/res/layout/activity_cadastro.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.auth.CadastroActivity">
+    tools:context=".auth.CadastroActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/nav_host_fragment"
