+++ mobile_hivemind/app/src/main/res/layout/activity_login.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/background"
     android:padding="24dp"
-    tools:context=".ui.auth.login.LoginActivity">
+    tools:context=".auth.login.LoginActivity">
 
     <TextView
         android:id="@+id/textView"
