+++ mobile_hivemind/app/src/main/res/layout/fragment_cadastro_endereco.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background"
-    tools:context=".ui.auth.cadastro.CadastroEnderecoFragment">
+    tools:context=".auth.cadastro.CadastroEnderecoFragment">
 
     <TextView
         android:id="@+id/txtContinuar"
