+++ mobile_hivemind/app/src/main/res/layout/fragment_criacao_conta.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background"
-    tools:context=".ui.auth.cadastro.CriacaoContaFragment">
+    tools:context=".auth.cadastro.CriacaoContaFragment">
 
     <!-- Título -->
     <TextView
