+++ mobile_hivemind/app/src/main/res/layout/fragment_verificacao_cell.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background"
-    tools:context=".ui.auth.VerificacaoCellFragment" >
+    tools:context=".auth.VerificacaoCellFragment" >
 
 
     <Button
