+++ mobile_hivemind/app/src/main/res/layout/fragment_verificacao_email.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background"
-    tools:context=".ui.auth.VerificacaoEmailFragment">
+    tools:context=".auth.VerificacaoEmailFragment">
 
     <Button
         android:id="@+id/btnVoltar"
